Cleanups & code splitting
Splitted SlateJS editor code to several files.
